Use test util
